update history page components
