+++ client/src/app/_components/HistoryTable.tsx
@@ -1,0 +1,128 @@
+"use client";
+
+import React, { useState } from "react";
+import {
+  Table as ReactTable,
+  useReactTable,
+  getCoreRowModel,
+  getFilteredRowModel,
+  getPaginationRowModel,
+  ColumnDef,
+  flexRender,
+  SortingState,
+  getSortedRowModel,
+} from "@tanstack/react-table";
+import { ArrowUpDown, ArrowDownUp } from "lucide-react";
+
+import { HistoryFilter } from "~/app/_components/HistoryFilter";
+import { ScreenshotData } from "~/lib/types";
+
+export function HistoryTable({
+  data,
+  columns,
+  setStatusFilter,
+}: {
+  data: ScreenshotData[];
+  columns: ColumnDef<ScreenshotData>[];
+  setStatusFilter: (statusFilter: string) => void;
+}) {
+  const [sorting, setSorting] = useState<SortingState>([]);
+
+  const table = useReactTable({
+    data,
+    columns,
+    state: {
+      sorting,
+    },
+    onSortingChange: setSorting,
+    getCoreRowModel: getCoreRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    debugTable: true,
+  });
+
+  return (
+    <div className="w-full">
+      <HistoryFilter table={table} setStatusFilter={setStatusFilter} />
+      <table className="w-full border-2">
+        <thead>
+          {table.getHeaderGroups().map((headerGroup) => (
+            <tr key={headerGroup.id} className="border-b-2">
+              {headerGroup.headers.map((header) => {
+                return (
+                  <th
+                    key={header.id}
+                    colSpan={header.colSpan}
+                    className="px-4 py-6"
+                  >
+                    {header.isPlaceholder ? null : (
+                      <div
+                        {...{
+                          className: header.column.getCanSort()
+                            ? "cursor-pointer select-none flex"
+                            : "",
+                          onClick: header.column.getToggleSortingHandler(),
+                        }}
+                      >
+                        {flexRender(
+                          header.column.columnDef.header,
+                          header.getContext(),
+                        )}
+                        {{
+                          asc: <ArrowDownUp />,
+                          desc: <ArrowUpDown />,
+                        }[header.column.getIsSorted() as string] ?? null}
+                      </div>
+                    )}
+                  </th>
+                );
+              })}
+            </tr>
+          ))}
+        </thead>
+        <tbody>
+          {table.getRowModel().rows.map((row) => {
+            return (
+              <tr key={row.id} className="border-b-2">
+                {row.getVisibleCells().map((cell) => {
+                  return (
+                    <td key={cell.id} className="truncate... px-4 py-4">
+                      {flexRender(
+                        cell.column.columnDef.cell,
+                        cell.getContext(),
+                      )}
+                    </td>
+                  );
+                })}
+              </tr>
+            );
+          })}
+        </tbody>
+      </table>
+      <div className="mt-8 flex items-center gap-2">
+        <button
+          className="rounded-md border-2 bg-blue-500 p-2 px-6 text-white"
+          onClick={() => table.previousPage()}
+          disabled={!table.getCanPreviousPage()}
+        >
+          Previous
+        </button>
+        <button
+          className="rounded-md border-2 bg-blue-500 p-2 px-6 text-white"
+          onClick={() => table.nextPage()}
+          disabled={!table.getCanNextPage()}
+        >
+          Next
+        </button>
+        <span className="flex items-center gap-1">
+          <div>Page</div>
+          <strong>
+            {table.getState().pagination.pageIndex + 1} of{" "}
+            {table.getPageCount()}
+          </strong>
+        </span>
+      </div>
+    </div>
+  );
+}
